Verify schnorr signatures only
 * Remove schnorr module

 * Add XOnly implementation in pubkey

 * Add SchnorrSignature implementation to the signature
