+++ Cargo.toml
@@ -16,7 +16,6 @@ std = [
   "light-bitcoin-primitives/std",
   "light-bitcoin-script/std",
   "light-bitcoin-serialization/std",
-  "light-bitcoin-schnorr/std"
 ]
 derive = ["light-bitcoin-serialization/derive"]
 
@@ -28,7 +27,6 @@ light-bitcoin-merkle = { path = "merkle", default-features = false }
 light-bitcoin-primitives = { path = "primitives", default-features = false }
 light-bitcoin-script = { path = "script", default-features = false }
 light-bitcoin-serialization = { path = "serialization", default-features = false }
-light-bitcoin-schnorr = { path = "schnorr", default-features = false }
 
 [workspace]
 members = [
@@ -40,5 +38,4 @@ members = [
   "script",
   "serialization",
   "serialization-derive",
-  "schnorr"
 ]
