+++ keys/Cargo.toml
@@ -13,6 +13,7 @@ std = [
   "hex/std",
   "libsecp256k1/std",
   "serde",
+  "sha2/std",
 
   "light-bitcoin-crypto/std",
   "light-bitcoin-primitives/std",
@@ -25,6 +26,8 @@ codec = { package = "parity-scale-codec", version = "2.0.0",  default-features =
 hex = { version = "0.4", default-features = false }
 libsecp256k1 = { version = "0.3.5", default-features = false, features = ["hmac"] }
 serde = { version = "1.0", features = ["derive"], optional = true }
+digest = { version = "0.9.0", default-features = false }
+sha2 = { version = "0.9.5", default-features = false }
 
 light-bitcoin-crypto = { path = "../crypto", default-features = false }
 light-bitcoin-primitives = { path = "../primitives", default-features = false }
