+++ keys/src/public.rs
@@ -1,4 +1,7 @@
-use core::{fmt, ops};
+use core::{
+    convert::{TryFrom, TryInto},
+    fmt, ops,
+};
 
 use light_bitcoin_crypto::dhash160;
 use light_bitcoin_primitives::{H264, H512, H520};
@@ -7,9 +10,13 @@ use codec::{Decode, Encode};
 #[cfg(feature = "std")]
 use serde::{Deserialize, Serialize};
 
-use crate::error::Error;
-use crate::signature::{CompactSignature, Signature};
-use crate::{AddressHash, Message};
+use crate::{
+    error::Error,
+    schnorr::verify_schnorr,
+    signature::{CompactSignature, SchnorrSignature, Signature},
+    tagged::HashInto,
+    AddressHash, Message,
+};
 
 /// Secret public key
 #[derive(Ord, PartialOrd, Eq, PartialEq, Copy, Clone)]
@@ -84,6 +91,18 @@ impl Public {
         Ok(secp256k1::verify(&message, &signature, &public))
     }
 
+    pub fn verify_schnorr(&self, message: &Message, signature: [u8; 64]) -> Result<bool, Error> {
+        let public = match self {
+            Public::Normal(pubkey) => secp256k1::PublicKey::parse(pubkey.as_fixed_bytes())?,
+            Public::Compressed(pubkey) => {
+                secp256k1::PublicKey::parse_compressed(pubkey.as_fixed_bytes())?
+            }
+        };
+        let xonly = public.try_into()?;
+        let signature = SchnorrSignature::try_from(signature)?;
+        verify_schnorr(&signature, message, xonly)
+    }
+
     pub fn verify_compact(&self, message: &Message, signature: &[u8; 64]) -> Result<bool, Error> {
         let public = match self {
             Public::Normal(pubkey) => secp256k1::PublicKey::parse(pubkey.as_fixed_bytes())?,
@@ -113,6 +132,128 @@ impl Public {
             Public::Normal(public)
         };
         Ok(public)
+    }
+}
+
+/// An [`XOnly`] is the compressed representation of a [`PublicKey`] which
+/// only stores the x-coordinate of the point.
+///
+/// X-only public keys become equivalent to a compressed public key
+/// that is the X-only key prefixed by the byte 0x02
+#[derive(Debug, Clone, PartialEq, Copy, Eq)]
+pub struct XOnly(pub [u8; 32]);
+
+impl XOnly {
+    pub fn on_curve(&self) -> Result<bool, Error> {
+        let mut elem = secp256k1::curve::Field::default();
+        let mut affine_coords = secp256k1::curve::Affine::default();
+        if elem.set_b32(&self.0) && affine_coords.set_xquad(&elem) {
+            Ok(true)
+        } else {
+            Err(Error::XCoordinateNotExist)
+        }
+    }
+}
+
+/// Convert [`Field`] to [`XOnly`]
+impl From<&mut secp256k1::curve::Field> for XOnly {
+    fn from(field: &mut secp256k1::curve::Field) -> Self {
+        field.normalize();
+        let slice = field.b32();
+        Self(slice)
+    }
+}
+
+impl TryFrom<&[u8]> for XOnly {
+    type Error = Error;
+
+    fn try_from(slice: &[u8]) -> Result<Self, Self::Error> {
+        let mut bytes = [0u8; 32];
+        bytes.copy_from_slice(slice);
+        Ok(Self(bytes))
+    }
+}
+
+/// Parse [`XOnly`] from hex
+impl TryFrom<&str> for XOnly {
+    type Error = Error;
+
+    fn try_from(value: &str) -> Result<Self, Self::Error> {
+        let x_bytes = hex::decode(value)?;
+        if x_bytes.len() != 32 {
+            return Err(Error::InvalidStringLength);
+        }
+        x_bytes[..].try_into()
+    }
+}
+
+/// Parse [`XOnly`] from 32 bytes
+impl TryFrom<[u8; 32]> for XOnly {
+    type Error = Error;
+
+    fn try_from(value: [u8; 32]) -> Result<Self, Self::Error> {
+        let mut elem = secp256k1::curve::Field::default();
+        let mut affine_coords = secp256k1::curve::Affine::default();
+        if elem.set_b32(&value) && affine_coords.set_xquad(&elem) {
+            Ok(Self(value))
+        } else {
+            Err(Error::XCoordinateNotExist)
+        }
+    }
+}
+
+/// Convert [`PublicKey`] to [`XOnly`]
+///
+/// The public keys constructed by such an algorithm (assuming they use the 33-byte compressed encoding)
+/// need to be converted by dropping the first byte.
+///
+/// [BIP340]: https://github.com/bitcoin/bips/blob/master/bip-0340.mediawiki#public-key-conversion
+impl TryFrom<secp256k1::PublicKey> for XOnly {
+    type Error = Error;
+
+    fn try_from(pubkey: secp256k1::PublicKey) -> Result<Self, Self::Error> {
+        let conmpressed = pubkey.serialize_compressed();
+        let (_, right) = conmpressed.split_at(1);
+        right.try_into()
+    }
+}
+
+/// Convert [`XOnly`] to [`PublicKey`]
+impl TryInto<secp256k1::PublicKey> for XOnly {
+    type Error = Error;
+
+    fn try_into(self) -> Result<secp256k1::PublicKey, Self::Error> {
+        let mut x = secp256k1::curve::Field::default();
+        let _ = x.set_b32(&self.0);
+        x.normalize();
+
+        let mut elem = secp256k1::curve::Affine::default();
+        elem.set_xquad(&x);
+        elem.y.normalize();
+        // determine the first byte of the compressed format public key
+        let tag = if elem.y.is_odd() {
+            // need to convert y to an even number
+            secp256k1::util::TAG_PUBKEY_EVEN
+        } else {
+            secp256k1::util::TAG_PUBKEY_ODD
+        };
+        // construct compressed public key
+        let mut c = [0u8; 33];
+        for (i, byte) in c.iter_mut().enumerate() {
+            if i == 0 {
+                *byte = tag;
+                continue;
+            }
+            *byte = x.b32()[i - 1];
+        }
+        let p = secp256k1::PublicKey::parse_compressed(&c)?;
+        Ok(p)
+    }
+}
+
+impl HashInto for XOnly {
+    fn hash_into(&self, hash: &mut impl digest::Digest) {
+        hash.update(self.0)
     }
 }
